+++ main.go
@@ -6,27 +6,6 @@ import (
 	"strings"
 )
 
-func buildImbalancedTransactionMsg(
-	filePath string,
-	lineNumber int,
-	amounts map[string]Amount,
-) string {
-	amountStrs := []string{}
-	for currency, amount := range amounts {
-		amountAndCurrency := fmt.Sprintf("%v %v", amount, currency)
-		amountStrs = append(amountStrs, amountAndCurrency)
-	}
-	imbalancedTransactionMsg := "%v:%v imbalanced transaction, (total amount) = %v"
-	// FIXME: 1000 USD + -1800 JPY
-	msg := fmt.Sprintf(
-		imbalancedTransactionMsg,
-		filePath,
-		lineNumber,
-		strings.Join(amountStrs, " + "),
-	)
-	return msg
-}
-
 func isZeroAmount(amounts map[string]Amount) bool {
 	for _, v := range amounts {
 		if v != 0 {
@@ -40,14 +19,12 @@ func main() {
 	var filePath = flag.String("f", "", "ledger/hledger transaction file")
 	flag.Parse()
 
-	fileContent, _ := readFileContent(*filePath)
-
-	knownAccountsStr, _ := readFileContent("fixtures/accounts.txt")
+	knownAccountsStr, _ := readFileContent("fixtures/accounts.txt") // FIXME: error handling
 	knownAccounts := strings.Split(knownAccountsStr, "\n")
 
 	countNewlines := 1
-	transactionStrs := strings.Split(fileContent, "\n\n")
-	for _, transactionStr := range transactionStrs {
+	transactionsStr, _ := readFileContent(*filePath) // FIXME: error handling
+	for _, transactionStr := range strings.Split(transactionsStr, "\n\n") {
 		_, transaction := parseTransactionStr(transactionStr)
 
 		// Check the transaction is balanced or not
